+++ lib/read/lib/Cardano/Wallet/Read/Block/Gen/BlockParameters.hs
@@ -10,7 +10,7 @@ module Cardano.Wallet.Read.Block.Gen.BlockParameters
     )
 where
 
-import Cardano.Wallet.Read.Block.BlockNo
+import Cardano.Read.Ledger.Block.BlockNo
     ( BlockNo (..)
     )
 import Cardano.Wallet.Read.Block.SlotNo
