+++ lib/read/lib/Cardano/Wallet/Read/Block/Gen/Byron.hs
@@ -25,7 +25,7 @@ import Cardano.Crypto.ProtocolMagic
 import Cardano.Crypto.Signing
     ( SigningKey (..)
     )
-import Cardano.Wallet.Read.Block.BlockNo
+import Cardano.Read.Ledger.Block.BlockNo
     ( BlockNo (..)
     )
 import Cardano.Wallet.Read.Block.Gen.BlockParameters
