+++ lib/read/lib/Cardano/Wallet/Read/Block/Gen/Shelley.hs
@@ -75,7 +75,7 @@ import Cardano.Wallet.Read
     , TxT
     , unTx
     )
-import Cardano.Wallet.Read.Block.BlockNo
+import Cardano.Read.Ledger.Block.BlockNo
     ( BlockNo (..)
     )
 import Cardano.Wallet.Read.Block.Gen.BlockParameters
